Removed finalizer and changed Dispose method to be used by the base finalizer. SqliteConnection now holds a list of weak references to the commands, since the prepared statements are now completely disposed by the command object itself.
# Conflicts:
#	src/Microsoft.Data.Sqlite.Core/SqliteConnection.cs
